整理查询方式，方便后面重构.
