+++ mybatis-plus-generator/src/main/java/com/baomidou/mybatisplus/generator/query/AbstractDatabaseQuery.java
@@ -1,0 +1,47 @@
+/*
+ * Copyright (c) 2011-2021, baomidou (jobob@qq.com).
+ * <p>
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ * <p>
+ * https://www.apache.org/licenses/LICENSE-2.0
+ * <p>
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+package com.baomidou.mybatisplus.generator.query;
+
+import com.baomidou.mybatisplus.generator.config.DataSourceConfig;
+import com.baomidou.mybatisplus.generator.config.builder.ConfigBuilder;
+import org.jetbrains.annotations.NotNull;
+
+
+/**
+ * @author nieqiurong
+ */
+public abstract class AbstractDatabaseQuery implements IDatabaseQuery {
+
+    protected final ConfigBuilder configBuilder;
+
+    protected final DataSourceConfig dataSourceConfig;
+
+    public AbstractDatabaseQuery(@NotNull ConfigBuilder configBuilder) {
+        this.configBuilder = configBuilder;
+        this.dataSourceConfig = configBuilder.getDataSourceConfig();
+    }
+
+    @NotNull
+    public ConfigBuilder getConfigBuilder() {
+        return configBuilder;
+    }
+
+    @NotNull
+    public DataSourceConfig getDataSourceConfig() {
+        return dataSourceConfig;
+    }
+
+}
